Addressed review comments. Refactored code. Created the IODescriptorDownloadable type, separated download tests from the desciptors.
