+++ python/tank/descriptor/descriptor.py
@@ -69,15 +69,15 @@ def create_descriptor(
 
     # use the environment variable if set - if not, fall back on the override or default locations
     if os.environ.get(constants.BUNDLE_CACHE_PATH_ENV_VAR):
-        bundle_cache_root_override = os.path.expandvars(
-            os.path.expanduser(os.environ.get(constants.BUNDLE_CACHE_PATH_ENV_VAR))
+        bundle_cache_root_override = os.path.expanduser(
+            os.path.expandvars(os.environ.get(constants.BUNDLE_CACHE_PATH_ENV_VAR))
         )
     elif bundle_cache_root_override is None:
         bundle_cache_root_override = _get_default_bundle_cache_root()
         filesystem.ensure_folder_exists(bundle_cache_root_override)
     else:
         # expand environment variables
-        bundle_cache_root_override = os.path.expandvars(os.path.expanduser(bundle_cache_root_override))
+        bundle_cache_root_override = os.path.expanduser(os.path.expandvars(bundle_cache_root_override))
 
     fallback_roots = fallback_roots or []
