+++ python/tank/descriptor/io_descriptor/git.py
@@ -11,9 +11,8 @@ import os
 import uuid
 import shutil
 import tempfile
-import subprocess
 
-from .base import IODescriptorBase
+from .downloadable import IODescriptorDownloadable
 from ... import LogManager
 from ...util.process import subprocess_check_output, SubprocessCalledProcessError
 
@@ -22,6 +21,7 @@ from ...util import filesystem
 
 log = LogManager.get_logger(__name__)
 
+
 class TankGitError(TankError):
     """
     Errors related to git communication
@@ -29,7 +29,7 @@ class TankGitError(TankError):
     pass
 
 
-class IODescriptorGit(IODescriptorBase):
+class IODescriptorGit(IODescriptorDownloadable):
     """
     Base class for git descriptors.
 
@@ -87,6 +87,7 @@ class IODescriptorGit(IODescriptorBase):
         """
         # ensure *parent* folder exists
         parent_folder = os.path.dirname(target_path)
+
         filesystem.ensure_folder_exists(parent_folder)
 
         # first probe to check that git exists in our PATH
