+++ tests/descriptor_tests/test_git.py
@@ -8,17 +8,13 @@
 # agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
 # not expressly granted therein are reserved by Shotgun Software Inc.
 
-from functools import reduce
-import multiprocessing
 import os
-import sys
-import time
 
 import sgtk
 from sgtk.descriptor import Descriptor
-from tank_test.tank_test_base import *
+from tank_test.tank_test_base import setUpModule # noqa
+from tank_test.tank_test_base import TankTestBase, skip_if_git_missing
 
-from tank_test.tank_test_base import TankTestBase, skip_if_git_missing, temp_env_var
 
 class TestGitIODescriptor(TankTestBase):
     """
@@ -121,7 +117,6 @@ class TestGitIODescriptor(TankTestBase):
         latest_desc.copy(copy_target)
         self.assertTrue(os.path.exists(os.path.join(copy_target, ".git")))
 
-
     @skip_if_git_missing
     def test_branch_shorthash(self):
 
@@ -142,7 +137,6 @@ class TestGitIODescriptor(TankTestBase):
             desc.get_path(),
             os.path.join(self.bundle_cache, "gitbranch", "tk-config-default.git", "3e6a681")
         )
-
 
     @skip_if_git_missing
     def test_branch(self):
@@ -211,85 +205,3 @@ class TestGitIODescriptor(TankTestBase):
         copy_target = os.path.join(self.project_root, "test_copy_target")
         latest_desc.copy(copy_target)
         self.assertTrue(os.path.exists(os.path.join(copy_target, ".git")))
-
-    @skip_if_git_missing
-    def test_downloads_to_bundle_cache(self):
-        """
-        Tests local downloads to the bundle cache for git descriptors.
-        """
-
-        # skip this test on windows or py2.5 where multiprocessing isn't available
-        if sys.platform == "win32" or sys.version_info < (2, 6):
-            return
-
-        def _download_bundles(target=None):
-            """
-            :param target: The path to which the bundle is to be downloaded.
-            """
-            location_dict_tag = {
-                "type": "git",
-                "path": self.git_repo_uri,
-                "version": "v0.15.0"
-            }
-            location_dict_branch = {
-                "type": "git_branch",
-                "path": self.git_repo_uri,
-                "branch": "33014_nuke_studio"
-            }
-            if target:
-                with temp_env_var(SHOTGUN_BUNDLE_CACHE_PATH=target):
-                    desc_git_tag = self._create_desc(location_dict_tag)
-                    desc_git_branch = self._create_desc(location_dict_branch, True)
-            else:
-                desc_git_tag = self._create_desc(location_dict_tag)
-                desc_git_branch = self._create_desc(location_dict_branch, True)
-            desc_git_tag.download_local()
-            desc_git_branch.download_local()
-
-        processes = []
-        errors = []
-
-        # test concurrent downloads to a shared bundle cache by multiple processes.
-
-        # the shared bundle cache path to which git data is to be downloaded.
-        shared_dir = os.path.join(self.tank_temp, "shared_bundle_cache")
-        try:
-            # spawn 10 processes that begin downloading data to the shared path.
-            for x in range(10):
-                process = multiprocessing.Process(target=_download_bundles, args=(shared_dir,))
-                process.start()
-                processes.append(process)
-        except Exception as e:
-            errors.append(e)
-
-        # wait until all processes have finished
-        all_processes_finished = False
-        while not all_processes_finished:
-            time.sleep(0.1)
-            sys.stderr.write(".")
-            all_processes_finished = all(not (process.is_alive()) for process in processes)
-
-        # Make sure the number of processes forked are as expected.
-        self.assertEqual(len(processes), 10, "Failed to spawn the expected number of processes.")
-
-        # Make sure the required files/folders exist on disk
-        self.assertTrue(os.path.exists(
-            os.path.join(self.tank_temp, "shared_bundle_cache", "git", "tk-config-default.git", "v0.15.0")
-        ), "Failed to find the shared bundle cache directory for the git descriptor on disk.")
-
-        self.assertTrue(os.path.exists(
-            os.path.join(self.tank_temp, "shared_bundle_cache", "gitbranch", "tk-config-default.git", "e1c03fa")
-        ), "Failed to find the shared bundle cache directory for the git descriptor on disk.")
-
-        # bit-wise OR the exit codes of all processes.
-        all_processes_exit_code = reduce(
-            lambda x, y: x | y,
-            [proc.exitcode for proc in processes]
-        )
-
-        # Make sure none of the child processes had non-zero exit statuses.
-        self.assertEqual(
-            all_processes_exit_code,
-            0,
-            "Failed to write concurrently to shared bundle cache: %s" % ",".join(errors)
-        )
